textual-showcase/pyproject.toml
[tool.poetry]
name = "rp-textual"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
textual-dev = "^1.6.1"
ipython = "^8.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
